% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{calc.het}
\alias{calc.het}
\title{Calculate Age-related Expression and Heterogeneity Changes}
\usage{
calc.het(exprmat, age, age_in = "days", age_to = "pw-0.25",
  batch_corr = "NC", modex = "linear", tr_log2 = T,
  sc_features = T, covariates = NA, het_change_met = "spearman",
  padj_met = "fdr")
}
\arguments{
\item{exprmat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}

\item{age}{a numeric vector, where the names correspond to samples (the same as colnames of the given matrix).}

\item{age_in}{type of the age vector, allowed values are days or years. defaults to 'days'}

\item{age_to}{final format of age vector. allowed values are 'years', 'days', 'pw-N', and 'lg-N', where N is any number. 'pw' means power. e.g. pw-0.5 means sqrt(age), and 'lg' means log, e.g. lg-2 means log2(age).defaults to 'pw-0.25'}

\item{batch_corr}{batch correction strategy. available values are i) NC: No Correction, ii) QN: Quantile Normalization, iii) LR: Linear regression (requires covariates), iv) SVA: surrogate variable analysis, v) LR+QN: Linear regression followed by quantile normalization, and vi) SVA+QN: SVA followed by quantile normalization. Defaults to 'NC'.}

\item{modex}{expression change calculation method. 'linear' or 'loess', defaults to 'linear'}

\item{tr_log2}{logical to set log2 transforming expression matrix. defaults to TRUE.}

\item{sc_features}{logical to set whether to scale features. defaults to TRUE.}

\item{covariates}{a list of covarietes where each element is a vector with sampleIDs as names}

\item{het_change_met}{heterogeneity change calculation method. 'LR', for 'linear regression', or and correlation method accepted by cor.test() function.}

\item{padj_met}{method for multiple test correction. value is passed to 'p.adjust' function. defaults to 'fdr'.}
}
\value{
a list object with summary results including expression level and heterogeneity changes, input values, intermediate values, and heterogeneity matrix.
}
\description{
Calculate Age-related Expression and Heterogeneity Changes
}
