% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cfr_quantileNorm}
\alias{cfr_quantileNorm}
\title{Quantile Normalize Expression Values}
\usage{
cfr_quantileNorm(mat)
}
\arguments{
\item{mat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}
}
\value{
quantile normalized expression values as a matrix
}
\description{
A function to apply quantile normalization. It is build on preprocessCore::normalize.quantiles function, but keeps the dimnames as is.
}
\examples{
myexp <- sapply( 1:10, function(i){
rnorm(n = 10000, mean = sample(1:3, 1), sd = sample(c(1, 3), 1))
})
myexp_qn <- cfr_quantileNorm(myexp)
par(mfrow = c(1,2))
boxplot(myexp, main = 'Before QN')
boxplot(myexp_qn, main = 'After QN')

}
