% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trans_log2}
\alias{trans_log2}
\title{Log2 Transformation}
\usage{
trans_log2(mat)
}
\arguments{
\item{mat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}
}
\value{
log2 transformed matrix
}
\description{
Log2 Transformation
}
\examples{
myexp <- sapply(1:10, function(i) rnbinom(10000, mu = 4, size = 1))
myexp_l2 <- trans_log2(myexp)
par(mfrow = c(1,2))
hist(myexp)
hist(myexp_l2)

}
