% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cfr_linReg}
\alias{cfr_linReg}
\title{Removing Confounding Factor Effects using Linear Regression}
\usage{
cfr_linReg(mat, cov)
}
\arguments{
\item{mat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}

\item{cov}{a list with covariates to regress out from the expression matrix}
}
\value{
a list with three elements, i) corrected expression matrix, ii) coefficients for the covariates and iii) p values for each covariate
}
\description{
Removing Confounding Factor Effects using Linear Regression
}
