% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cfr_sva}
\alias{cfr_sva}
\title{Removing Confounding Factor Effects using SVA}
\usage{
cfr_sva(mat, age, cov = NA)
}
\arguments{
\item{mat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}

\item{age}{a numeric vector, where the names correspond to samples (the same as colnames of the given matrix).}

\item{cov}{covariates to analyse SVs.}
}
\value{
a list with i) a numeric matrix with the expression values, ii) SVs for each sample, and iii) correlation between SVs and covariates
}
\description{
Removing Confounding Factor Effects using SVA
}
