% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{feature_scale}
\alias{feature_scale}
\title{Scale Expression Values for Features (probsets, transcripts, genes)}
\usage{
feature_scale(mat)
}
\arguments{
\item{mat}{a numeric matrix with the expression values, where columns are the samples and rows are probesets, transcripts, or genes.}
}
\value{
scaled matrix
}
\description{
Scale Expression Values for Features (probsets, transcripts, genes)
}
\examples{
myexp <- t(sapply( 1:10000, function(i){ sample((0:12000)/1000,100) }))
myexp_scaled <- feature_scale(myexp)
par(mfrow=c(1,2))
boxplot(t(myexp[1:10,]), main='original')
boxplot(t(myexp_scaled[1:10,]), main = 'scaled')

}
